GitHub repositories

Search for GitHub repositories using specific queries.

github_repositories(
  query,
  token = gh::gh_token(),
  .limit = Inf,
  verbose = TRUE
)

Source

GitHub Docs: Search repositories

GitHub Docs: Constructing a search query

Searching GitHub repos

Case-sensitive GitHub searches

Examples of using gh to search repositories

Arguments

query

Regex query to filter files by.

token

GitHub Personal Authentication Token (PAT). See here for further instructions.

.limit

Number of records to return. This can be used instead of manual pagination. By default it is NULL, which means that the defaults of the GitHub API are used. You can set it to a number to request more (or less) records, and also to Inf to request all records. Note, that if you request many records, then multiple GitHub API calls are used to get them, and this can take a potentially long time.

verbose

Print messages.

Examples

if (FALSE) { # \dontrun{
repos <- github_repositories(query="language:r", .limit=100)
} # }